+++ uni_modules/uni-file-picker/components/uni-file-picker/uni-file-picker.vue
@@ -40,7 +40,7 @@
 </template>
 
 <script>
-
+	import chooseAndUploadFile from './choose-and-upload-file.js'
 	import uploadImage from './upload-image.vue'
 	import uploadFile from './upload-file.vue'
 	let fileInput = null
@@ -192,6 +192,13 @@ export default {
 					}
 					newFils.push(files ? files : v)
 				})
+				let data  = null
+				if (this.returnType === 'object') {
+					data = this.backObject(newFils)[0]
+				} else {
+					data = this.backObject(newFils)
+				}
+				this.formItem && this.formItem.setValue(data)
 				this.files = newFils
 			},
 			immediate: true
@@ -239,11 +246,35 @@ export default {
 		}
 	},
 	created() {
-		// this.files = Object.assign([], this.value)
+		// TODO 兼容不开通服务空间的情况
+		if(!(uniCloud.config && uniCloud.config.provider)){
+			this.noSpace = true
+			uniCloud.chooseAndUploadFile = chooseAndUploadFile
+		}
 		this.tempData = {}
-
+		this.form = this.getForm('uniForms')
+		this.formItem = this.getForm('uniFormsItem')
+		if (this.form && this.formItem) {
+			if (this.formItem.name) {
+				this.rename = this.formItem.name
+				this.form.inputChildrens.push(this)
+			}
+		}
 	},
 	methods: {
+		/**
+		 * 获取父元素实例
+		 */
+		getForm(name = 'uniForms') {
+			let parent = this.$parent;
+			let parentName = parent.$options.name;
+			while (parentName !== name) {
+				parent = parent.$parent;
+				if (!parent) return false;
+				parentName = parent.$options.name;
+			}
+			return parent;
+		},
 		/**
 		 * 继续上传
 		 */
@@ -270,6 +301,7 @@ export default {
 		 * 选择文件
 		 */
 		choose() {
+			
 			if (this.disabled) return
 			if (this.files.length >= Number(this.limitLength) && this.showType !== 'grid' && this.returnType === 'array') {
 				uni.showToast({
@@ -334,6 +366,7 @@ export default {
 							if (this.limitLength - this.files.length <= 0) break
 							files[i].uuid = Date.now()
 							let filedata = await this.getFileData(files[i], this.fileMediatype)
+							filedata.file = files[i]
 							filedata.progress = 0
 							filedata.status = 'ready'
 							this.files.push(filedata)
@@ -345,7 +378,7 @@ export default {
 						})
 						res.tempFiles = files
 						// 停止自动上传
-						if (!this.autoUpload) {
+						if (!this.autoUpload || this.noSpace) {
 							res.tempFiles = []
 							// TODO 先放弃这个实现 ，不能全部上传
 							// return new Promise((resolve) => {
@@ -402,13 +435,14 @@ export default {
 					this.files[index].url = item.path
 					this.files[index].status = 'error'
 					this.files[index].errMsg = item.errMsg
-					this.files[index].progress = -1
+					// this.files[index].progress = -1
 					errorData.push(this.files[index])
 					errorTempFilePath.push(this.files[index].url)
 				} else {
 					this.files[index].errMsg = ''
 					this.files[index].url = item.url
 					this.files[index].status = 'success'
+					this.files[index].progress += 1
 					successData.push(this.files[index])
 					tempFilePath.push(this.files[index].url)
 				}
@@ -446,7 +480,8 @@ export default {
 				idx = this.files.findIndex(p => p.uuid === progressEvent.tempFile.uuid)
 			}
 			if (idx === -1 || !this.files[idx]) return
-			this.files[idx].progress = percentCompleted
+			// fix by mehaotian 100 就会消失，-1 是为了让进度条消失
+			this.files[idx].progress = percentCompleted - 1
 			// 上传中
 			this.$emit('progress', {
 				index: idx,
@@ -460,8 +495,10 @@ export default {
 		 * @param {Object} index
 		 */
 		delFile(index) {
-			let fileData = this.files[index]
-			this.$emit('delete', fileData)
+			this.$emit('delete', {
+				tempFile: this.files[index],
+				tempFilePath: this.files[index].url
+			})
 			this.files.splice(index, 1)
 			this.$nextTick(()=>{
 				this.setEmit()
@@ -511,6 +548,7 @@ export default {
 				uuid: files.uuid,
 				extname: extname || '',
 				cloudPath: files.cloudPath,
+				file:files.file,
 				fileType: files.fileType,
 				url: files.path || files.path,
 				size: files.size, //单位是字节
@@ -520,9 +558,12 @@ export default {
 			}
 			if (type === 'image') {
 				const imageinfo = await this.getFileInfo(files.path)
+				delete filedata.video
 				filedata.image.width = imageinfo.width
 				filedata.image.height = imageinfo.height
 				filedata.image.location = imageinfo.path
+			}else{
+				delete filedata.image
 			}
 			return filedata
 		},